meshes stacked on top of each other always merge a light regardless of ini settings, (if 1 is shadow merged light will be shadow), shadow lights receive a boost  to min dist required of 82 centemeters

+++ include/LightData.h
@@ -52,13 +52,13 @@ struct Overlay
 
 struct LightData {
 
-	static std::map<uint64_t, LightConfig> configIDToJsonCfg;
+	static std::map<uint32_t, LightConfig> configIDToJsonCfg;
 
 	static std::unordered_map<std::string, std::vector<LightConfig>> nodeNameToJsonCfg;
 
 	static std::unordered_map<std::string, std::vector<LightConfig>> meshPathToJsonCfg;
 
-	static std::unordered_map<uint64_t, LightConfig> defaultConfigs;
+	static std::unordered_map<uint32_t, LightConfig> defaultConfigs;
 
 	static bool excludeLightEditorID(const RE::TESObjectLIGH* light);
